* Split protocols into smaller protocols and files. * Split DataSourceController protocol conformance into separate files. * Split SectionDataModel and SectionDataController into separate files.
