+++ DataSourceController/Protocols/Cells/CellView.swift
@@ -1,0 +1,7 @@
+//  Copyright © 2019 Pedro Daniel Prieto Martínez. All rights reserved.
+
+import Foundation
+
+public protocol CellView {
+    func configure(with dataController: CellDataController)
+}
